учел рекомендации ревью
