+++ FakeNFT/Scenes /Profle/FavoriteNFTScreen/FavoriteNFTViewModel.swift
@@ -91,7 +91,7 @@ final class FavoriteNFTViewModel {
                 if let index = self.likedNFTs.firstIndex(
                     where: { $0.id == nft.id }
                 ) {
-                    self.likedNFTs[index].name = self.extractNFTName(
+                    self.likedNFTs[index].name = nftImageUrl.extractNFTName(
                         from: nftImageUrl
                     ) ?? nft.originalName
                 }
@@ -110,34 +110,6 @@ final class FavoriteNFTViewModel {
                 )
             }
         }
-    }
-    
-    func extractNFTName(
-        from urlString: String
-    ) -> String? {
-        let pattern = #"\/([^\/]+)\/\d+\.png$"#
-        let regex = try? NSRegularExpression(
-            pattern: pattern,
-            options: []
-        )
-        let nsString = urlString as NSString
-        let results = regex?.firstMatch(
-            in: urlString,
-            options: [],
-            range: NSRange(
-                location: 0,
-                length: nsString.length
-            )
-        )
-        
-        if let range = results?.range(
-            at: 1
-        ) {
-            return nsString.substring(
-                with: range
-            )
-        }
-        return nil
     }
     
     func ratingImage(
