+++ FakeNFT/Scenes /Profle/ProfileService/ProfileByRequest.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct ProfileRequest: NetworkRequest {
     var endpoint: URL? {
-        URL(string: "\(RequestConstants.baseURL)/api/v1/profile/1")
+        URL(string: "\(RequestConstants.baseURL)\(RequestConstants.profilePath)")
     }
     var dto: Dto?
 }
