+++ FakeNFT/Scenes /Profle/ProfileService/ProfilePutRequest.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct ProfilePutRequest: NetworkRequest {
    var endpoint: URL? {
-       URL(string: "\(RequestConstants.baseURL)/api/v1/profile/1")
+       URL(string: "\(RequestConstants.baseURL)\(RequestConstants.profilePath)")
    }
    var httpMethod: HttpMethod = .put
    var dto: Dto?
