+++ FakeNFT/Services/Requests/RequestConstants.swift
@@ -2,4 +2,5 @@ enum RequestConstants {
     static let baseURL = "https://d5dn3j2ouj72b0ejucbl.apigw.yandexcloud.net"
     #warning("Instert your token here")
     static let token = ""
+    static let profilePath = "/api/v1/profile/1"
 }
